+++ tests/onnx2c_refs/test_onnx03_conv2d_tiny.c
@@ -24,29 +24,29 @@ static const float weight_Conv_out_bias[1] = {
 };
 
 void node1_conv2d_f32(const float input[1][1][5][5], const float weight[1][1][3][3], const float bias[1], float out[1][1][3][3]) {
-    int64_t in_per_group = 1 / 1;
-    int64_t out_per_group = 1 / 1;
-    int64_t out_pad_h = 0;
-    int64_t out_pad_w = 0;
+    ssize_t in_per_group = 1 / 1;
+    ssize_t out_per_group = 1 / 1;
+    ssize_t out_pad_h = 0;
+    ssize_t out_pad_w = 0;
     (void)out_pad_h;
     (void)out_pad_w;
     for (ssize_t n = 0; n < 1; ++n) {
         for (ssize_t oc = 0; oc < 1; ++oc) {
-            int64_t group = (int64_t)oc / out_per_group;
+            ssize_t group = (ssize_t)oc / out_per_group;
             for (ssize_t oh = 0; oh < 3; ++oh) {
                 for (ssize_t ow = 0; ow < 3; ++ow) {
                     float acc = 0.0f;
-                    int64_t in_h_base = (int64_t)oh * 1 - 0;
-                    int64_t in_w_base = (int64_t)ow * 1 - 0;
+                    ssize_t in_h_base = (ssize_t)oh * 1 - 0;
+                    ssize_t in_w_base = (ssize_t)ow * 1 - 0;
                     for (ssize_t ic = 0; ic < in_per_group; ++ic) {
-                        int64_t in_c = group * in_per_group + (int64_t)ic;
+                        ssize_t in_c = group * in_per_group + (ssize_t)ic;
                         for (ssize_t kh = 0; kh < 3; ++kh) {
-                            int64_t in_h_idx = in_h_base + (int64_t)kh * 1;
+                            ssize_t in_h_idx = in_h_base + (ssize_t)kh * 1;
                             if (in_h_idx < 0 || in_h_idx >= 5) {
                                 continue;
                             }
                             for (ssize_t kw = 0; kw < 3; ++kw) {
-                                int64_t in_w_idx = in_w_base + (int64_t)kw * 1;
+                                ssize_t in_w_idx = in_w_base + (ssize_t)kw * 1;
                                 if (in_w_idx < 0 || in_w_idx >= 5) {
                                     continue;
                                 }
